Extract base IntegrationTest class from commands integration tests
